pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
license = "MIT"
name = "open-french-law-rag-pipeline"
readme = "README.md"
version = "0.0.1"

[tool.poetry.dependencies]
chromadb = "^0.4.18"
click = "^8.1.7"
datasets = "^2.15.0"
langchain = "^0.0.341"
litellm = "^1.9.4"
openai = "^1.3.6"
pandas = "^2.1.3"
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
sentence-transformers = "^2.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
